requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
cymem==2.0.7
d6tcollect==1.0.6
d6tflow==0.2.6
distlib==0.3.6
docutils==0.19
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
filelock==3.11.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
gast==0.4.0
google-api-core==2.11.0
google-api-python-client==2.85.0
google-api-python-client-py3==1.2
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-api-gateway==1.7.1
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
h5py==3.8.0
httplib2==0.22.0
huggingface-hub==0.13.4
idna==3.4
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
langcodes==3.3.0
libclang==16.0.0
lockfile==0.12.2
luigi==3.2.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
multidict==6.0.4
murmurhash==1.0.9
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.4
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.0
pathy==0.10.1
pdfminer==20191125
pdfminer.six==20221105
Pillow==9.5.0
platformdirs==3.2.0
polars==0.17.8
preshed==3.0.8
proto-plus==1.22.2
protobuf==4.22.3
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
PyMuPDF==1.22.0
pyparsing==3.0.9
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
scrapetube==2.3.1
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post4
smart-open==6.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-text==2.12.1
termcolor==2.2.0
thinc==8.1.9
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.3
tornado==6.3
tqdm==4.65.0
transformers==4.28.1
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
wasabi==1.1.1
Werkzeug==2.2.3
wrapt==1.14.1
yarl==1.8.2
youtube-transcript-api==0.6.0
